move iter to models
