+++ rest/reference/splits.go
@@ -4,13 +4,12 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/polygon-io/client-go/rest/client"
 	"github.com/polygon-io/client-go/rest/models"
 )
 
 // ListSplitsIter is an iterator for the ListSplits method.
 type ListSplitsIter struct {
-	client.Iter
+	models.Iter
 }
 
 // Split returns the current result that the iterator points to.
@@ -43,7 +42,7 @@ func (c *Client) ListSplits(ctx context.Context, params *models.ListSplitsParams
 	}
 
 	return &ListSplitsIter{
-		Iter: client.NewIter(ctx, uri, func(uri string) (models.ListResponse, []interface{}, error) {
+		Iter: models.NewIter(ctx, uri, func(uri string) (models.ListResponse, []interface{}, error) {
 			res := &models.ListSplitsResponse{}
 			err := c.CallURL(ctx, http.MethodGet, uri, res, options...)
